+++ packages/service/src/index.ts
@@ -1,4 +1,4 @@
-import { db } from "@acme/db/client";
+import { db } from "@acme/db-backbone/client";
 
 import type { PostService } from "./post";
 import { createPostService } from "./post";