+++ packages/service/src/post.ts
@@ -1,9 +1,9 @@
 import type { InferSelectModel } from "drizzle-orm";
 
-import type { Database } from "@acme/db/client";
+import type { Database } from "@acme/db-backbone/client";
 import type { CreatePostInput } from "@acme/validators";
-import { desc, eq } from "@acme/db";
-import { Post } from "@acme/db/schema";
+import { desc, eq } from "@acme/db-backbone";
+import { Post } from "@acme/db-backbone/schema";
 
 type PostRow = InferSelectModel<typeof Post>;